Refactor common code in entry point classes (for xpcli and xp cmd) to base class
